cloud_storage: Add translation unit for test/util.h
The header is currently only used in one translation unit. To use it in
other tests outside of that translation unit the corresponding .cc file
has to be added and the code has to be moved there.
